Create resources directory and provide default world file
This world file is copied from the default world supplied by the Gazebo
simulator. However, this file has a comment with a double hyphenation in
it which is not supported by the Python xml parser. Therefore, we copied
the file and removed the comment to solve the problem. This also ensures
the default path for the `--base` argument resolves correctly when not
specified.