Refactor ardupilot implemenatation to support start-up sequencing
The ArduPilot firmware requires knowledge of the gazebo container in
order to set the url for the gazebo plugin. To accomplish this, the
gazebo component is started first, and then the container name extracted
and passed to the ArduPilot firmware component.

We also drop the JointGazeboFirmwareComponent extensions and for
module-specific implementations because the inheritance creates issues
when trying to return different types and this functionality is really
only used in the ArduPilot module. The PX4 module does not need the same
information, so it can be implemented more simply.