Refactor ComponentContainer options into independent base class
Since the ReporterNode implementation does not inherit from the
ContainerNode class, the ReporterComponent cannot inherit from the
ContainerComponent class because the `start` method for the
ContainerComponent and its derivates must return a ContainerNode
or derivative. However, both the ContainerComponent and the
ReporterComponent share many constructor arguments and it is beneficial
to have the `attrs` library generate the constructor for us. Therefore,
we define a common base class that both ContainerComponent and
ReporterComponent inherit from that defines the common attributes, and
allows us to override them while still retaining the constructor
generation.